feat: moved types to outer package
